pyproject.toml
[tool.poetry]
name = "vegetables-for-you"
version = "0.0.1"
description = "Get vegetables information"
authors = ["Lin-Rexter <aaaa35963596@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "vegetables_for_you"}]

[tool.poetry.dependencies]
python = "^3.9"


[tool.poetry.group.dev.dependencies]
line-bot-sdk = "2.3.0"
beautifulsoup4 = "4.11.1"
psycopg = "3.1.1"
numpy = "1.23.3"
Scrapy = "2.6.2"
Flask = "^2.2.2"
gunicorn = "^20.1.0"
black = "^22.8.0"
bandit = "^1.7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
